insert/requirements.txt
python-dotenv
influxdb-client